Fixed plotting issue
Previously, during plotting pandas dataframes had 'object' dtypes which caused the matplotlib axes formatter to plt them as categories rather than numbers. This has been fixed by casting these dataframes and series as float dtypes.

+++ empyric/graphics.py
@@ -160,9 +160,9 @@ class Plotter:
             if x == 'Time':
                 xdata = self.full_data.index
             else:
-                xdata = self.full_data[x]
+                xdata = self.full_data[x].astype(float)
 
-            ydata = self.full_data[ys]
+            ydata = self.full_data[ys].astype(float)
 
         if ax.lines and not errorbars:
             # For simple plots (i.e. not error bar plots), if already plotted,
